Extract LLM policy loss to llm_policy_loss module
Move the pure-tensor llm_policy_loss_fn and _k3_kl into a new Liger-free
module (agilerl.algorithms.core.llm_policy_loss) so the math is
unit-testable without liger-kernel. fused_llm_policy_loss now imports
that function, enforces HAS_LIGER_KERNEL at import time and raises a
clear ImportError when liger-kernel is missing. Update tests and
docstrings accordingly.
