+++ tests/test_algorithms/test_fused_llm_policy_loss.py
@@ -1,8 +1,10 @@
-"""Tests for ``agilerl.algorithms.core.fused_llm_policy_loss``.
+"""Tests for ``agilerl.algorithms.core.llm_policy_loss`` and
+``agilerl.algorithms.core.fused_llm_policy_loss``.
 
-The math in :func:`llm_policy_loss_fn` is testable without ``liger_kernel``
-installed. End-to-end autograd-Function tests live in the GPU-box
-benchmarks since Liger is a CUDA-leaning dependency.
+The math in :func:`llm_policy_loss_fn` lives in the Liger-free
+``llm_policy_loss`` module and is testable everywhere. End-to-end
+autograd-Function tests are gated on ``HAS_LIGER_KERNEL`` since the
+fused module fails fast at import time without ``liger-kernel``.
 """
 
 from __future__ import annotations
@@ -13,37 +15,28 @@ import sys
 import pytest
 import torch
 
-from agilerl.algorithms.core.fused_llm_policy_loss import (
+from agilerl.algorithms.core.llm_policy_loss import (
     _k3_kl,
     llm_policy_loss_fn,
 )
 
 
-def test_no_liger_fallback_resolves_module_with_none_function():
-    """When ``HAS_LIGER_KERNEL=False``, the module must still import
-    cleanly: the ``LigerFusedLinearPPOBase`` name is stubbed to ``object``
-    so the class definition resolves, then the public symbol
-    ``LigerFusedLinearPolicyLossFunction`` is reassigned to ``None`` so
-    callers' ``is None`` guard fires. This branch is unreachable on
-    Linux CI (Liger is installed) and is the only platform-fallback
-    path coverage source — exercise it explicitly via ``importlib.reload``,
-    which preserves module identity (no stale ``sys.modules`` entries to
-    confuse later tests that imported the original module).
+def test_no_liger_fused_module_raises_import_error(monkeypatch):
+    """``fused_llm_policy_loss`` requires ``liger-kernel`` and fails fast
+    at import time without it. Exercise the guard by forcing
+    ``HAS_LIGER_KERNEL=False`` and triggering a fresh import. Works on
+    both Liger-present (Linux CI) and Liger-absent (macOS dev) machines:
+    a failed module body never enters ``sys.modules``, so the cached
+    real module — if any — survives untouched.
     """
     import agilerl
-    import agilerl.algorithms.core.fused_llm_policy_loss as mod
 
-    original_has_liger = agilerl.HAS_LIGER_KERNEL
-    try:
-        agilerl.HAS_LIGER_KERNEL = False
-        importlib.reload(mod)
-        assert mod.LigerFusedLinearPolicyLossFunction is None
-        # ``LigerFusedLinearPPOBase`` is the stub object base used so the
-        # class body can be defined without Liger installed.
-        assert mod.LigerFusedLinearPPOBase is object
-    finally:
-        agilerl.HAS_LIGER_KERNEL = original_has_liger
-        importlib.reload(mod)
+    monkeypatch.setattr(agilerl, "HAS_LIGER_KERNEL", False)
+    mod_name = "agilerl.algorithms.core.fused_llm_policy_loss"
+    monkeypatch.delitem(sys.modules, mod_name, raising=False)
+
+    with pytest.raises(ImportError, match="liger-kernel"):
+        importlib.import_module(mod_name)
 
 
 @pytest.mark.parametrize(
@@ -601,14 +594,17 @@ class TestLlmPpoLossFnTurnMode:
 
 # ---------------------------------------------------------------------------
 # Autograd Function tests — only run when liger_kernel is importable.
-# Without Liger, ``LigerFusedLinearPolicyLossFunction`` resolves to ``None`` and
-# the math fn above already covers the loss formula.
+# ``fused_llm_policy_loss`` raises at import time without Liger, so gate
+# the import on ``HAS_LIGER_KERNEL``; the math fn above already covers
+# the loss formula on no-Liger platforms.
 # ---------------------------------------------------------------------------
 
 from agilerl import HAS_LIGER_KERNEL  # noqa: E402
-from agilerl.algorithms.core.fused_llm_policy_loss import (  # noqa: E402
-    LigerFusedLinearPolicyLossFunction,
-)
+
+if HAS_LIGER_KERNEL:
+    from agilerl.algorithms.core.fused_llm_policy_loss import (  # noqa: E402
+        LigerFusedLinearPolicyLossFunction,
+    )
 
 
 @pytest.mark.skipif(
